Hack so it works. Must fix.
Under limited cases, the previous code worked, but it doesn't when it dealed
with several components in series in a feedback loop. An easy hack to fix it is
to evaluate *twice* the nodes between clocks, but it is both inefficient and
incorrect: an input that reacts to updates will change its outputs and thus
mess up the feedback loops.

We must analyse the graph and update the nodes in the correct order. The
analysis' result could be cached in the form of pointers in sigprocs that forms
chains (a embedded single-linked list) that is followed when evaluating nodes in
the graph.